% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{match_pipeline}
\alias{match_pipeline}
\title{match pipeline}
\usage{
match_pipeline(fields)
}
\arguments{
\item{fields}{data.frame or JSON of field & values to include in the match
pipeline. A data.frame must be one row, and where the column names are the
keys and the row contains the values.}
}
\value{
match pipline string
}
\description{
builds the \code{$match} pipeline
}

