% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{project_pipeline}
\alias{project_pipeline}
\title{project pipeline}
\usage{
project_pipeline(fields, exclude_id = TRUE)
}
\arguments{
\item{fields}{vector of fields to include in the project pipeline}

\item{exclude_id}{logical}
}
\value{
project pipeline string
}
\description{
builds the \code{$project} pipeline
}

