% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{unwind_pipeline}
\alias{unwind_pipeline}
\title{undwind pipeline}
\usage{
unwind_pipeline(field)
}
\arguments{
\item{field}{string indicating the array field to unwind.}
}
\value{
unwind pipeline string
}
\description{
builds the \code{$unwind} pipeline
}

